made adjustments in features

+++ src/components/cart/CartItem.js
@@ -6,6 +6,12 @@ import { Link } from "react-router-dom/cjs/react-router-dom.min";
 
 const CartItem = ({ cartItems, dispatch }) => {
   const handleRemoveItemFromCart = (id) => {
+    const permission = window.confirm(
+      "Are you sure you want to delete this product?"
+    );
+    if (!permission) {
+      return;
+    }
     dispatch(removeFromCart(id));
   };
